Lazy logic refactoring
