+++ Sources/Extensions/CollectionViewManager+UICollectionViewDataSource.swift
@@ -7,7 +7,7 @@ import UIKit.UICollectionView
 extension CollectionViewManager: UICollectionViewDataSource {
 
     open func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        sectionItemsProvider.numberOfItems(inSection: section)
+        sectionItemsWrapper.numberOfCellItems(inSection: section)
     }
 
     open func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
@@ -18,7 +18,7 @@ extension CollectionViewManager: UICollectionViewDataSource {
     }
 
     open func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return sectionItemsProvider.numberOfSections
+        return sectionItemsWrapper.numberOfSections
     }
 
     open func collectionView(_ collectionView: UICollectionView,
