+++ Sources/Extensions/CollectionViewManager+UICollectionViewDelegateFlowLayout.swift
@@ -9,31 +9,31 @@ extension CollectionViewManager: UICollectionViewDelegateFlowLayout {
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              sizeForItemAt indexPath: IndexPath) -> CGSize {
-        sectionItemsProvider.sizeForCellItem(at: indexPath, in: collectionView)
+        sectionItemsWrapper.sizeForCellItem(at: indexPath, in: collectionView)
     }
 
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              insetForSectionAt section: Int) -> UIEdgeInsets {
-        return sectionItemsProvider[section]?.insets ?? .zero
+        return sectionItemsWrapper[section]?.insets ?? .zero
     }
 
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return sectionItemsProvider[section]?.minimumLineSpacing ?? 0
+        return sectionItemsWrapper[section]?.minimumLineSpacing ?? 0
     }
 
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return sectionItemsProvider[section]?.minimumInteritemSpacing ?? 0
+        return sectionItemsWrapper[section]?.minimumInteritemSpacing ?? 0
     }
 
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              referenceSizeForHeaderInSection section: Int) -> CGSize {
-        let optionalItem = sectionItemsProvider[section]?.reusableViewItems.first { reusableViewItem in
+        let optionalItem = sectionItemsWrapper[section]?.reusableViewItems.first { reusableViewItem in
             reusableViewItem.type == .header
         }
         guard let item = optionalItem else {
@@ -45,7 +45,7 @@ extension CollectionViewManager: UICollectionViewDelegateFlowLayout {
     open func collectionView(_ collectionView: UICollectionView,
                              layout collectionViewLayout: UICollectionViewLayout,
                              referenceSizeForFooterInSection section: Int) -> CGSize {
-        let optionalItem = sectionItemsProvider[section]?.reusableViewItems.first { reusableViewItem in
+        let optionalItem = sectionItemsWrapper[section]?.reusableViewItems.first { reusableViewItem in
             reusableViewItem.type == .footer
         }
         guard let item = optionalItem else {
